requirements.txt
# Core dependencies
schedule==1.2.0
lxml==5.1.0
pymongo==4.6.3
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.5

# Development dependencies (optional)
# pytest==7.4.0
# black==23.7.0
# flake8==6.0.0
# isort==5.12.0